Update wishlist logic & Add newsletter app

+++ wishlist/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
+from django.http import JsonResponse
 
 from products.models import Product
 from profiles.models import UserProfile
@@ -18,24 +19,31 @@ def wishlist(request):
 
 @login_required
 def add_to_wishlist(request, product_id):
-    """ Add a product to the user's wishlist. """
+    """A view to add an item to the wishlist"""
     product = get_object_or_404(Product, pk=product_id)
     user_profile = UserProfile.objects.get(user=request.user)
 
     if Wishlist.objects.filter(user_profile=user_profile, product=product).exists():
         messages.warning(request, f'{product.name} is already in your Wishlist.')
     else:
-        wishlist_item = Wishlist.objects.create(user_profile=user_profile, product=product)
-        messages.success(request, f'{wishlist_item.product.name} added to Wishlist successfully!')
+        Wishlist.objects.create(user_profile=user_profile, product=product)
+        messages.success(request, f'{product.name} added to Wishlist successfully!')
 
+    # Redirect back to the product detail page or another suitable page
     return redirect(reverse('product_detail', args=[product_id]))
+
 
 @login_required
 def remove_from_wishlist(request, product_id):
-    """ Remove item from Wishlist when remove icon is clicked """
+    """ A view to remove an item from the wishlist """
     product = get_object_or_404(Product, pk=product_id)
     user_profile = UserProfile.objects.get(user=request.user)
-    wishlist_item = get_object_or_404(Wishlist, user_profile=user_profile, product=product)
-    wishlist_item.delete()
-    messages.success(request, f'{product.name} has been successfully removed.')
+    wishlist_item = Wishlist.objects.filter(user_profile=user_profile, product=product).first()
+
+    if wishlist_item:
+        wishlist_item.delete()
+        messages.success(request, f'{product.name} has been successfully removed.')
+    else:
+        messages.warning(request, f'{product.name} was not found in your Wishlist.')
+
     return redirect(reverse('wishlist'))
